requirements.txt
pyyaml
pexpect
crayons
